fixed lds generation
